Revert "Solved dirent bug, unsigned int -> unsigned long long for pos"
This reverts commit 4fe30c01a66b9a4f0781f793d302edf663f2f958.

+++ include/client/open_dir.hpp
@@ -41,7 +41,7 @@ class OpenDir: public OpenFile {
     public:
         OpenDir(const std::string& path);
         void add(const std::string& name, const FileType& type);
-        const DirEntry& getdent(unsigned long long pos);
+        const DirEntry& getdent(unsigned int pos);
         size_t size();
 };
